@@ -21,23 +21,22 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.action.exception.ActionException;
-
 import org.openqa.selenium.By;
 
 /**
  * {@link Custom} is a subclass of {@link MultiPhaseAction} which performs custom action by phases
  * to avoid blocking the execution of other browser actions.
- * 
+ *
  * <p>There are two constructors of {@link Custom}:
+ *
  * <ol>
- * <li>{@link Custom#Custom(java.util.function.Function, java.util.List) } is used to make
- * {@link Custom} act as a {@link MultiPhaseAction}, therefore {@link MultiPhased#noNextPhase()}
- * must be called explicitly before ending custom action.</li>
- * <li>{@link Custom#Custom(java.util.function.Consumer, java.util.List) } is used to make
- * {@link Custom} act as a {@link SinglePhaseAction}.</li>
+ *   <li>{@link Custom#Custom(java.util.function.Function, java.util.List) } is used to make {@link
+ *       Custom} act as a {@link MultiPhaseAction}, therefore {@link MultiPhased#noNextPhase()} must
+ *       be called explicitly before ending custom action.
+ *   <li>{@link Custom#Custom(java.util.function.Consumer, java.util.List) } is used to make {@link
+ *       Custom} act as a {@link SinglePhaseAction}.
  * </ol>
  *
  * @author Kimberly
@@ -49,23 +48,23 @@ public class Custom extends MultiPhaseAction {
 
   /**
    * Create a {@link Custom} acting like a {@link MultiPhaseAction}.
-   * 
+   *
    * @param customAction multiple-phased custom action
-   */    
+   */
   public Custom(Function<MultiPhased, Consumer<ActionComposer>> customAction) {
     this(customAction, null);
   }
 
   /**
    * Create a {@link Custom} acting like a {@link MultiPhaseAction}.
-   * 
+   *
    * @param customAction multiple-phased custom action
    * @param frameBySequence the sequence of the frame locating mechanism for the frame where the
    *     custom action to be performed against
-   */    
-  public Custom(Function<MultiPhased, Consumer<ActionComposer>> customAction,
-      List<By> frameBySequence) {
-    this.customFunc = customAction.apply((MultiPhased)this);
+   */
+  public Custom(
+      Function<MultiPhased, Consumer<ActionComposer>> customAction, List<By> frameBySequence) {
+    this.customFunc = customAction.apply((MultiPhased) this);
     if (frameBySequence != null) {
       this.frameBySequence.addAll(frameBySequence);
     }
@@ -74,7 +73,7 @@ public class Custom extends MultiPhaseAction {
 
   /**
    * Create a {@link Custom} acting like a {@link SinglePhaseAction}.
-   * 
+   *
    * @param customAction custom action
    */
   public Custom(Consumer<ActionComposer> customAction) {
@@ -83,7 +82,7 @@ public class Custom extends MultiPhaseAction {
 
   /**
    * Create a {@link Custom} acting like a {@link SinglePhaseAction}.
-   * 
+   *
    * @param customAction custom action
    * @param frameBySequence the sequence of the frame locating mechanism for the frame where the
    *     custom action to be performed against
@@ -100,9 +99,9 @@ public class Custom extends MultiPhaseAction {
   protected void performMultiPhase() {
     ActionComposer actionComposer = this.getComposer();
     try {
-      //firefox doesn't switch focus to top after switch to window,
-      //so recovery step is required
-      switchToTopForFirefox(); 
+      // firefox doesn't switch focus to top after switch to window,
+      // so recovery step is required
+      switchToTopForFirefox();
       actionComposer.switchToInnerFrame(this.frameBySequence);
       this.customFunc.accept(actionComposer);
       if (this.actAsSinglePhase) {
@@ -116,8 +115,11 @@ public class Custom extends MultiPhaseAction {
 
   @Override
   public String toString() {
-    return String.format("%s(%s):%s", Custom.class.getSimpleName(),
-        String.valueOf(actAsSinglePhase),  String.join(",",frameBySequence.stream().map(
-            s -> s.toString()).collect(Collectors.toList())));
+    return String.format(
+        "%s(%s):%s",
+        Custom.class.getSimpleName(),
+        String.valueOf(actAsSinglePhase),
+        String.join(
+            ",", frameBySequence.stream().map(s -> s.toString()).collect(Collectors.toList())));
   }
 }
