@@ -22,21 +22,18 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.action.exception.ActionException;
-
 import org.openqa.selenium.By;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebElement;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * {@link Extract} is a subclass of {@link MultiPhaseAction} which extract information from element
  * by phases to avoid blocking the execution of other browser actions.
- * 
+ *
  * @author Kimberly
  */
 public class Extract extends MultiPhaseAction {
@@ -49,11 +46,11 @@ public class Extract extends MultiPhaseAction {
 
   /**
    * Create an action to extract information from element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param textVariableName text variable name; non-empty name means to get the visible (i.e. not
-   *     hidden by CSS) text of the element(including sub-elements) as a variable of
-   *     {@link ActionComposer}
+   *     hidden by CSS) text of the element(including sub-elements) as a variable of {@link
+   *     ActionComposer}
    */
   public Extract(By by, String textVariableName) {
     this(by, null, textVariableName, null);
@@ -61,13 +58,13 @@ public class Extract extends MultiPhaseAction {
 
   /**
    * Create an action to extract information from element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param frameBySequence the sequence of the frame locating mechanism for the element resides in
    *     frame(or frame in another frame and so on)
    * @param textVariableName text variable name; non-empty name means to get the visible (i.e. not
-   *     hidden by CSS) text of the element(including sub-elements) as a variable of
-   *     {@link ActionComposer}
+   *     hidden by CSS) text of the element(including sub-elements) as a variable of {@link
+   *     ActionComposer}
    */
   public Extract(By by, List<By> frameBySequence, String textVariableName) {
     this(by, frameBySequence, textVariableName, null);
@@ -75,12 +72,12 @@ public class Extract extends MultiPhaseAction {
 
   /**
    * Create an action to extract information from element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param attrVariableNames (attribute name, variable name) pairs to set as variables. For each
-   *     pair, {@link Extract} sets the property value of the element as a variable of
-   *     {@link ActionComposer} if the property exists. If property doesn't exists, {@link Extract}
-   *     sets the attribute value of the element as a variable of {@link ActionComposer}. If neither
+   *     pair, {@link Extract} sets the property value of the element as a variable of {@link
+   *     ActionComposer} if the property exists. If property doesn't exists, {@link Extract} sets
+   *     the attribute value of the element as a variable of {@link ActionComposer}. If neither
    *     exists, null-value variable is set.
    */
   public Extract(By by, Map<String, String> attrVariableNames) {
@@ -89,14 +86,14 @@ public class Extract extends MultiPhaseAction {
 
   /**
    * Create an action to extract information from element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param frameBySequence the sequence of the frame locating mechanism for the element resides in
    *     frame(or frame in another frame and so on)
    * @param attrVariableNames (attribute name, variable name) pairs to set as variables. For each
-   *     pair, {@link Extract} sets the property value of the element as a variable of
-   *     {@link ActionComposer} if the property exists. If property doesn't exists, {@link Extract}
-   *     sets the attribute value of the element as a variable of {@link ActionComposer}. If neither
+   *     pair, {@link Extract} sets the property value of the element as a variable of {@link
+   *     ActionComposer} if the property exists. If property doesn't exists, {@link Extract} sets
+   *     the attribute value of the element as a variable of {@link ActionComposer}. If neither
    *     exists, null-value variable is set.
    */
   public Extract(By by, List<By> frameBySequence, Map<String, String> attrVariableNames) {
@@ -105,16 +102,16 @@ public class Extract extends MultiPhaseAction {
 
   /**
    * Create an action to extract information from element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param textVariableName text variable name; non-empty name means to get the visible (i.e. not
-   *     hidden by CSS) text of the element(including sub-elements) as a variable of 
-   *     {@link ActionComposer}
+   *     hidden by CSS) text of the element(including sub-elements) as a variable of {@link
+   *     ActionComposer}
    * @param attrVariableNames (attribute name, variable name) pairs to set as variables. For each
-   *     pair, {@link Extract} sets the property value of the element as a variable of
-   *     {@link ActionComposer} if the property exists. If property doesn't exists, {@link Extract}
-   *     sets the attribute value of the element as a variable of {@link ActionComposer}. If
-   *     neither exists, null-value variable is set.
+   *     pair, {@link Extract} sets the property value of the element as a variable of {@link
+   *     ActionComposer} if the property exists. If property doesn't exists, {@link Extract} sets
+   *     the attribute value of the element as a variable of {@link ActionComposer}. If neither
+   *     exists, null-value variable is set.
    */
   public Extract(By by, String textVariableName, Map<String, String> attrVariableNames) {
     this(by, null, textVariableName, attrVariableNames);
@@ -122,20 +119,23 @@ public class Extract extends MultiPhaseAction {
 
   /**
    * Create an action to extract information from element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param frameBySequence the sequence of the frame locating mechanism for the element resides in
    *     frame(or frame in another frame and so on)
    * @param textVariableName text variable name; non-empty name means to get the visible (i.e. not
-   *     hidden by CSS) text of the element(including sub-elements) as a variable of
-   *     {@link ActionComposer}
+   *     hidden by CSS) text of the element(including sub-elements) as a variable of {@link
+   *     ActionComposer}
    * @param attrVariableNames (attribute name, variable name) pairs to set as variables. For each
-   *     pair, {@link Extract} sets the property value of the element as a variable of
-   *     {@link ActionComposer} if the property exists. If property doesn't exists, {@link Extract}
-   *     sets the attribute value of the element as a variable of {@link ActionComposer}. If
-   *     neither exists, null-value variable is set.
+   *     pair, {@link Extract} sets the property value of the element as a variable of {@link
+   *     ActionComposer} if the property exists. If property doesn't exists, {@link Extract} sets
+   *     the attribute value of the element as a variable of {@link ActionComposer}. If neither
+   *     exists, null-value variable is set.
    */
-  public Extract(By by, List<By> frameBySequence, String textVariableName,
+  public Extract(
+      By by,
+      List<By> frameBySequence,
+      String textVariableName,
       Map<String, String> attrVariableNames) {
     this.by = by;
     if (frameBySequence != null) {
@@ -151,27 +151,31 @@ public class Extract extends MultiPhaseAction {
   protected void performMultiPhase() {
     ActionComposer actionComposer = this.getComposer();
     try {
-      //firefox doesn't switch focus to top after switch to window,
-      //so recovery step is required
-      switchToTopForFirefox(); 
+      // firefox doesn't switch focus to top after switch to window,
+      // so recovery step is required
+      switchToTopForFirefox();
       actionComposer.switchToInnerFrame(this.frameBySequence);
       WebElement element = actionComposer.getWebDriver().findElement(this.by);
 
-      //get text when necessary
+      // get text when necessary
       if (!this.textVariableName.isEmpty()) {
         actionComposer.setVariable(this.textVariableName, element.getText());
       }
 
-      //get attribute when necessary
-      for (Map.Entry<String, String> entry: this.attrVariableNames.entrySet()) {
+      // get attribute when necessary
+      for (Map.Entry<String, String> entry : this.attrVariableNames.entrySet()) {
         actionComposer.setVariable(entry.getValue(), element.getAttribute(entry.getKey()));
       }
       noNextPhase();
     } catch (StaleElementReferenceException ignoreE) {
-      //with next phase when StaleElementReferenceException is encountered
+      // with next phase when StaleElementReferenceException is encountered
       if (LOGGER.isDebugEnabled()) {
-        LOGGER.debug("{}({}): encounter stale element:{}", ActionComposer.class.getSimpleName(),
-            actionComposer.getName(), toString(), ignoreE);
+        LOGGER.debug(
+            "{}({}): encounter stale element:{}",
+            ActionComposer.class.getSimpleName(),
+            actionComposer.getName(),
+            toString(),
+            ignoreE);
       }
     } catch (Exception e) {
       noNextPhase();
@@ -181,10 +185,19 @@ public class Extract extends MultiPhaseAction {
 
   @Override
   public String toString() {
-    return String.format("%s:%s/%s/%s/%s", Extract.class.getSimpleName(), by.toString(),
-        String.join(",",frameBySequence.stream().map(s -> s.toString())
-            .collect(Collectors.toList())), textVariableName,
-            String.join(",", attrVariableNames.entrySet().stream().map(
-                s -> s.getKey() + ":" + s.getValue()).collect(Collectors.toList())));
+    return String.format(
+        "%s:%s/%s/%s/%s",
+        Extract.class.getSimpleName(),
+        by.toString(),
+        String.join(
+            ",", frameBySequence.stream().map(s -> s.toString()).collect(Collectors.toList())),
+        textVariableName,
+        String.join(
+            ",",
+            attrVariableNames
+                .entrySet()
+                .stream()
+                .map(s -> s.getKey() + ":" + s.getValue())
+                .collect(Collectors.toList())));
   }
 }
