@@ -19,24 +19,20 @@ package org.kquiet.browser.action;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.action.exception.ActionException;
-
 import org.openqa.selenium.By;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * {@link MouseOver} is a subclass of {@link MultiPhaseAction} which moves the mouse to the middle
- * of an element.
- * {@link org.openqa.selenium.StaleElementReferenceException} may happen while {@link MouseOver}
- * tries to manipulate the element, so multi-phase is used to perform the action again.
- * 
+ * of an element. {@link org.openqa.selenium.StaleElementReferenceException} may happen while {@link
+ * MouseOver} tries to manipulate the element, so multi-phase is used to perform the action again.
+ *
  * @author Kimberly
  */
 public class MouseOver extends MultiPhaseAction {
@@ -47,7 +43,7 @@ public class MouseOver extends MultiPhaseAction {
 
   /**
    * Create an action representing moving mouse over some element.
-   * 
+   *
    * @param by the element locating mechanism
    */
   public MouseOver(By by) {
@@ -56,7 +52,7 @@ public class MouseOver extends MultiPhaseAction {
 
   /**
    * Create an action representing moving mouse over some element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param frameBySequence the sequence of the frame locating mechanism for the element resides in
    *     frame(or frame in another frame and so on)
@@ -72,17 +68,21 @@ public class MouseOver extends MultiPhaseAction {
   protected void performMultiPhase() {
     ActionComposer actionComposer = this.getComposer();
     try {
-      //firefox doesn't switch focus to top after switch to window, so recovery step is required
-      switchToTopForFirefox(); 
+      // firefox doesn't switch focus to top after switch to window, so recovery step is required
+      switchToTopForFirefox();
       actionComposer.switchToInnerFrame(this.frameBySequence);
       WebElement element = actionComposer.getWebDriver().findElement(this.by);
       new Actions(actionComposer.getWebDriver()).moveToElement(element).perform();
       noNextPhase();
     } catch (StaleElementReferenceException ignoreE) {
-      //with next phase when StaleElementReferenceException is encountered
+      // with next phase when StaleElementReferenceException is encountered
       if (LOGGER.isDebugEnabled()) {
-        LOGGER.debug("{}({}): encounter stale element:{}", ActionComposer.class.getSimpleName(),
-            actionComposer.getName(), toString(), ignoreE);
+        LOGGER.debug(
+            "{}({}): encounter stale element:{}",
+            ActionComposer.class.getSimpleName(),
+            actionComposer.getName(),
+            toString(),
+            ignoreE);
       }
     } catch (Exception e) {
       noNextPhase();
@@ -92,8 +92,11 @@ public class MouseOver extends MultiPhaseAction {
 
   @Override
   public String toString() {
-    return String.format("%s:%s/%s", MouseOver.class.getSimpleName(), by.toString(),
-        String.join(",",frameBySequence.stream().map(
-            s -> s.toString()).collect(Collectors.toList())));
+    return String.format(
+        "%s:%s/%s",
+        MouseOver.class.getSimpleName(),
+        by.toString(),
+        String.join(
+            ",", frameBySequence.stream().map(s -> s.toString()).collect(Collectors.toList())));
   }
 }
