@@ -19,24 +19,21 @@ package org.kquiet.browser.action;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.action.exception.ActionException;
-
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebElement;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * {@link ScrollToView} is a subclass of {@link MultiPhaseAction} which scrolls an element into
- * visible area of the browser window.
- * {@link org.openqa.selenium.StaleElementReferenceException} may happen while {@link ScrollToView}
- * tries to manipulate the element, so multi-phase is used to perform the action again.
- * 
+ * visible area of the browser window. {@link org.openqa.selenium.StaleElementReferenceException}
+ * may happen while {@link ScrollToView} tries to manipulate the element, so multi-phase is used to
+ * perform the action again.
+ *
  * @author Kimberly
  */
 public class ScrollToView extends MultiPhaseAction {
@@ -48,7 +45,7 @@ public class ScrollToView extends MultiPhaseAction {
 
   /**
    * Create an action to scroll to view a specified element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param toTop {@code true}: scroll to top;{@code false}: scroll to bottom
    */
@@ -58,7 +55,7 @@ public class ScrollToView extends MultiPhaseAction {
 
   /**
    * Create an action to scroll to view a specified element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param frameBySequence the sequence of the frame locating mechanism for the element resides in
    *     frame(or frame in another frame and so on)
@@ -76,18 +73,22 @@ public class ScrollToView extends MultiPhaseAction {
   protected void performMultiPhase() {
     ActionComposer actionComposer = this.getComposer();
     try {
-      //firefox doesn't switch focus to top after switch to window, so recovery step is required
-      switchToTopForFirefox(); 
+      // firefox doesn't switch focus to top after switch to window, so recovery step is required
+      switchToTopForFirefox();
       actionComposer.switchToInnerFrame(this.frameBySequence);
       WebElement element = actionComposer.getWebDriver().findElement(this.by);
       ((JavascriptExecutor) actionComposer.getWebDriver())
           .executeScript("arguments[0].scrollIntoView(arguments[1]);", element, this.toTop);
       noNextPhase();
     } catch (StaleElementReferenceException ignoreE) {
-      //with next phase when StaleElementReferenceException is encountered
+      // with next phase when StaleElementReferenceException is encountered
       if (LOGGER.isDebugEnabled()) {
-        LOGGER.debug("{}({}): encounter stale element:{}", ActionComposer.class.getSimpleName(),
-            actionComposer.getName(), toString(), ignoreE);
+        LOGGER.debug(
+            "{}({}): encounter stale element:{}",
+            ActionComposer.class.getSimpleName(),
+            actionComposer.getName(),
+            toString(),
+            ignoreE);
       }
     } catch (Exception e) {
       noNextPhase();
@@ -97,8 +98,12 @@ public class ScrollToView extends MultiPhaseAction {
 
   @Override
   public String toString() {
-    return String.format("%s:%s/%s/%s", ScrollToView.class.getSimpleName(), by.toString(),
-        String.join(",",frameBySequence.stream().map(
-            s -> s.toString()).collect(Collectors.toList())), String.valueOf(toTop));
+    return String.format(
+        "%s:%s/%s/%s",
+        ScrollToView.class.getSimpleName(),
+        by.toString(),
+        String.join(
+            ",", frameBySequence.stream().map(s -> s.toString()).collect(Collectors.toList())),
+        String.valueOf(toTop));
   }
 }
