@@ -19,23 +19,20 @@ package org.kquiet.browser.action;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.action.exception.ActionException;
-
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebElement;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * {@link SendKey} is a subclass of {@link MultiPhaseAction} which types into an element.
- * {@link org.openqa.selenium.StaleElementReferenceException} may happen while {@link Select} tries
- * to manipulate the element, so multi-phase is used to perform the action again.
- * 
+ * {@link SendKey} is a subclass of {@link MultiPhaseAction} which types into an element. {@link
+ * org.openqa.selenium.StaleElementReferenceException} may happen while {@link Select} tries to
+ * manipulate the element, so multi-phase is used to perform the action again.
+ *
  * @author Kimberly
  */
 public class SendKey extends MultiPhaseAction {
@@ -60,7 +57,7 @@ public class SendKey extends MultiPhaseAction {
 
   /**
    * Create an action to input keys to a specified element.
-   * 
+   *
    * @param by the element locating mechanism
    * @param frameBySequence the sequence of the frame locating mechanism for the element resides in
    *     frame(or frame in another frame and so on)
@@ -68,8 +65,8 @@ public class SendKey extends MultiPhaseAction {
    *     first
    * @param keysToSend character sequence to send to the element
    */
-  public SendKey(By by, List<By> frameBySequence, boolean clearBeforeSend,
-      CharSequence... keysToSend) {
+  public SendKey(
+      By by, List<By> frameBySequence, boolean clearBeforeSend, CharSequence... keysToSend) {
     this.by = by;
     if (frameBySequence != null) {
       this.frameBySequence.addAll(frameBySequence);
@@ -80,9 +77,9 @@ public class SendKey extends MultiPhaseAction {
 
   private CharSequence[] purifyCharSequences(CharSequence... sequence) {
     if (sequence == null) {
-      sequence = new CharSequence[]{};
+      sequence = new CharSequence[] {};
     } else {
-      for (int i = 0;i < sequence.length;i++) {
+      for (int i = 0; i < sequence.length; i++) {
         if (sequence[i] == null) {
           sequence[i] = "";
         }
@@ -93,15 +90,15 @@ public class SendKey extends MultiPhaseAction {
 
   /**
    * Clicks an element and then types into it.
-   * 
+   *
    * @param element the element to send keys to
    * @param clearBeforeSend {@code true}: clear before sending; {@code false}: send without clearing
    *     first
    * @param keysToSend character sequence to send to the element
    */
-  public static void clickToSendKeys(WebElement element, boolean clearBeforeSend,
-      CharSequence... keysToSend) {
-    //click before send key
+  public static void clickToSendKeys(
+      WebElement element, boolean clearBeforeSend, CharSequence... keysToSend) {
+    // click before send key
     element.click();
     if (clearBeforeSend) {
       element.clear();
@@ -116,17 +113,21 @@ public class SendKey extends MultiPhaseAction {
   protected void performMultiPhase() {
     ActionComposer actionComposer = this.getComposer();
     try {
-      //firefox doesn't switch focus to top after switch to window, so recovery step is required
-      switchToTopForFirefox(); 
+      // firefox doesn't switch focus to top after switch to window, so recovery step is required
+      switchToTopForFirefox();
       actionComposer.switchToInnerFrame(this.frameBySequence);
       WebElement element = actionComposer.getWebDriver().findElement(this.by);
       clickToSendKeys(element, this.clearBeforeSend, this.keysToSend);
       noNextPhase();
-    } catch (StaleElementReferenceException ignoreE) { 
-      //with next phase when StaleElementReferenceException is encountered
+    } catch (StaleElementReferenceException ignoreE) {
+      // with next phase when StaleElementReferenceException is encountered
       if (LOGGER.isDebugEnabled()) {
-        LOGGER.debug("{}({}): encounter stale element:{}", ActionComposer.class.getSimpleName(),
-            actionComposer.getName(), toString(), ignoreE);
+        LOGGER.debug(
+            "{}({}): encounter stale element:{}",
+            ActionComposer.class.getSimpleName(),
+            actionComposer.getName(),
+            toString(),
+            ignoreE);
       }
     } catch (Exception e) {
       noNextPhase();
@@ -136,8 +137,12 @@ public class SendKey extends MultiPhaseAction {
 
   @Override
   public String toString() {
-    return String.format("%s:%s/%s/%s", SendKey.class.getSimpleName(), by.toString(),
-        String.join(",", keysToSend), String.join(",",frameBySequence.stream().map(
-            s -> s.toString()).collect(Collectors.toList())));
+    return String.format(
+        "%s:%s/%s/%s",
+        SendKey.class.getSimpleName(),
+        by.toString(),
+        String.join(",", keysToSend),
+        String.join(
+            ",", frameBySequence.stream().map(s -> s.toString()).collect(Collectors.toList())));
   }
 }
