@@ -17,55 +17,51 @@
 package org.kquiet.browser.action;
 
 import java.util.List;
-
 import org.kquiet.browser.ActionComposer;
 
 /**
- * An interface which should be implemented by any action which delegates its execution to
- * {@link ActionComposer}.
- * 
+ * An interface which should be implemented by any action which delegates its execution to {@link
+ * ActionComposer}.
+ *
  * @author Kimberly
  */
 public interface Composable {
 
   /**
    * Get containing composer.
-   * 
+   *
    * @return containing composer
    */
   ActionComposer getComposer();
 
   /**
    * Set containing composer.
-   * 
+   *
    * @param composer the containing composer to set
    * @return invoking {@link Composable}
    */
   Composable setComposer(ActionComposer composer);
 
-  /**
-   * Perform action.
-   */
+  /** Perform action. */
   void perform();
 
   /**
    * Get the errors occurred during execution.
-   * 
+   *
    * @return the errors as a list of {@link Exception}
    */
   List<Exception> getErrors();
 
   /**
    * Check if this action is done.
-   * 
+   *
    * @return {@code true} if this action is done; {@code false} otherwise
    */
   boolean isDone();
 
-
   /**
    * Check if this action is failed.
-   * 
+   *
    * @return {@code true} if this action is failed; {@code false} otherwise
    */
   boolean isFail();
