@@ -21,11 +21,9 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Function;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.action.exception.ActionException;
 import org.kquiet.utility.Stopwatch;
-
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.TimeoutException;
@@ -35,21 +33,22 @@ import org.openqa.selenium.support.ui.FluentWait;
 /**
  * {@link WaitUntil} is a subclass of {@link MultiPhaseAction} which waits the evaluation result of
  * condition function by phases to avoid blocking the execution of other browser actions.
- * 
- * <p>{@link WaitUntil} waits until one of the following occurs:</p>
+ *
+ * <p>{@link WaitUntil} waits until one of the following occurs:
+ *
  * <ol>
- * <li>the condition function returns neither null nor false</li>
- * <li>the condition function throws an unignored exception</li>
- * <li>the timeout expires</li>
- * <li>the execution thread of this {@link WaitUntil} is interrupted</li>
+ *   <li>the condition function returns neither null nor false
+ *   <li>the condition function throws an unignored exception
+ *   <li>the timeout expires
+ *   <li>the execution thread of this {@link WaitUntil} is interrupted
  * </ol>
- * 
- * <p>When timeout expires, it will throw an
- * {@link org.kquiet.browser.action.exception.ActionException ActionException} if no timeout
- * callback function is supplied;
- * if a timeout callback function is supplied, it will execute the callback function instead of
- * throwing {@link org.kquiet.browser.action.exception.ActionException ActionException}.</p>
- * 
+ *
+ * <p>When timeout expires, it will throw an {@link
+ * org.kquiet.browser.action.exception.ActionException ActionException} if no timeout callback
+ * function is supplied; if a timeout callback function is supplied, it will execute the callback
+ * function instead of throwing {@link org.kquiet.browser.action.exception.ActionException
+ * ActionException}.
+ *
  * @author Kimberly
  * @param <V> the expected return type of condition function
  */
@@ -62,95 +61,121 @@ public class WaitUntil<V> extends MultiPhaseAction {
   private final int totalTimeout;
   private final int phaseTimeout;
   private final int pollInterval;
-  private final Function<WebDriver,V> conditionFunc;
+  private final Function<WebDriver, V> conditionFunc;
   private final Set<Class<? extends Throwable>> ignoreExceptions = new HashSet<>();
   private final Consumer<ActionComposer> timeoutCallback;
 
   /**
    * Create an action to wait with default phased timeout and polling interval until conditions are
    * met or timed out.
-   * 
+   *
    * @param conditionFunc the condition function for evaluation by phases
    * @param totalTimeout the maximum amount of time to wait totally
    */
-  public WaitUntil(Function<WebDriver,V> conditionFunc, int totalTimeout) {
+  public WaitUntil(Function<WebDriver, V> conditionFunc, int totalTimeout) {
     this(conditionFunc, totalTimeout, DEFAULT_PHASE_TIMEOUT, DEFAULT_POLL_INTERVAL, null, null);
   }
 
   /**
    * Create an action to wait until conditions are met or timed out.
-   * 
+   *
    * @param conditionFunc the condition function for evaluation by phases
    * @param totalTimeout the maximum amount of time to wait totally
    * @param phaseTimeout the maximum amount of time to wait for each execution phase
    * @param pollInterval how often the condition function should be evaluated(the cost of actually
    *     evaluating the condition function is not factored in)
    */
-  public WaitUntil(Function<WebDriver,V> conditionFunc, int totalTimeout, int phaseTimeout,
-      int pollInterval) {
+  public WaitUntil(
+      Function<WebDriver, V> conditionFunc, int totalTimeout, int phaseTimeout, int pollInterval) {
     this(conditionFunc, totalTimeout, phaseTimeout, pollInterval, null, null);
   }
 
   /**
    * Create an action to wait with default phased timeout and polling interval until conditions are
    * met or timed out.
-   * 
+   *
    * @param conditionFunc the condition function for evaluation by phases
    * @param totalTimeout the maximum amount of time to wait totally
    * @param timeoutCallback the callback function to be called when total timeout expires
    */
-  public WaitUntil(Function<WebDriver,V> conditionFunc, int totalTimeout,
+  public WaitUntil(
+      Function<WebDriver, V> conditionFunc,
+      int totalTimeout,
       Consumer<ActionComposer> timeoutCallback) {
-    this(conditionFunc, totalTimeout, DEFAULT_PHASE_TIMEOUT, DEFAULT_POLL_INTERVAL, null,
+    this(
+        conditionFunc,
+        totalTimeout,
+        DEFAULT_PHASE_TIMEOUT,
+        DEFAULT_POLL_INTERVAL,
+        null,
         timeoutCallback);
   }
 
   /**
    * Create an action to wait with default phased timeout and polling interval until conditions are
    * met or timed out.
-   * 
+   *
    * @param conditionFunc the condition function for evaluation by phases
    * @param totalTimeout the maximum amount of time to wait totally
    * @param ignoreExceptions the types of exceptions to ignore when evaluating condition function;
    */
-  public WaitUntil(Function<WebDriver,V> conditionFunc, int totalTimeout,
+  public WaitUntil(
+      Function<WebDriver, V> conditionFunc,
+      int totalTimeout,
       Set<Class<? extends Throwable>> ignoreExceptions) {
-    this(conditionFunc, totalTimeout, DEFAULT_PHASE_TIMEOUT, DEFAULT_POLL_INTERVAL,
-        ignoreExceptions, null);
+    this(
+        conditionFunc,
+        totalTimeout,
+        DEFAULT_PHASE_TIMEOUT,
+        DEFAULT_POLL_INTERVAL,
+        ignoreExceptions,
+        null);
   }
 
   /**
    * Create an action to wait with default phased timeout and polling interval until conditions are
    * met or timed out.
-   * 
+   *
    * @param conditionFunc the condition function for evaluation by phases
    * @param totalTimeout the maximum amount of time to wait totally
    * @param ignoreExceptions the types of exceptions to ignore when evaluating condition function;
    * @param timeoutCallback the callback function to be called when total timeout expires
    */
-  public WaitUntil(Function<WebDriver,V> conditionFunc, int totalTimeout,
-      Set<Class<? extends Throwable>> ignoreExceptions, Consumer<ActionComposer> timeoutCallback) {
-    this(conditionFunc, totalTimeout, DEFAULT_PHASE_TIMEOUT, DEFAULT_POLL_INTERVAL,
-        ignoreExceptions, timeoutCallback);
+  public WaitUntil(
+      Function<WebDriver, V> conditionFunc,
+      int totalTimeout,
+      Set<Class<? extends Throwable>> ignoreExceptions,
+      Consumer<ActionComposer> timeoutCallback) {
+    this(
+        conditionFunc,
+        totalTimeout,
+        DEFAULT_PHASE_TIMEOUT,
+        DEFAULT_POLL_INTERVAL,
+        ignoreExceptions,
+        timeoutCallback);
   }
 
   /**
    * Create an action to wait until conditions are met or timed out.
-   * 
+   *
    * @param conditionFunc the condition function for evaluation by phases
    * @param totalTimeout the maximum amount of time to wait totally
    * @param phaseTimeout the maximum amount of time to wait for each execution phase
    * @param pollInterval how often the condition function should be evaluated(the cost of actually
    *     evaluating the condition function is not factored in)
    * @param ignoreExceptions the types of exceptions to ignore when evaluating condition function.
-   *     If this parameter value is null or empty, then
-   *     {@link org.openqa.selenium.StaleElementReferenceException} and
-   *     {@link org.openqa.selenium.NoSuchElementException} are ignored; otherwise the given types
-   *     of exceptions are used.
+   *     If this parameter value is null or empty, then {@link
+   *     org.openqa.selenium.StaleElementReferenceException} and {@link
+   *     org.openqa.selenium.NoSuchElementException} are ignored; otherwise the given types of
+   *     exceptions are used.
    * @param timeoutCallback the callback function to be called when total timeout expires
    */
-  public WaitUntil(Function<WebDriver,V> conditionFunc, int totalTimeout, int phaseTimeout,
-      int pollInterval, Set<Class<? extends Throwable>> ignoreExceptions,
+  public WaitUntil(
+      Function<WebDriver, V> conditionFunc,
+      int totalTimeout,
+      int phaseTimeout,
+      int pollInterval,
+      Set<Class<? extends Throwable>> ignoreExceptions,
       Consumer<ActionComposer> timeoutCallback) {
     this.totalTimeout = totalTimeout;
     this.phaseTimeout = phaseTimeout;
@@ -172,7 +197,7 @@ public class WaitUntil<V> extends MultiPhaseAction {
       if (!actionComposer.switchToFocusWindow()) {
         throw new ActionException("can't switch to focus window");
       }
-      //firefox doesn't switch focus to top after switch to window, so recovery step is required
+      // firefox doesn't switch focus to top after switch to window, so recovery step is required
       switchToTopForFirefox();
       timeoutCallback.accept(actionComposer);
     } else {
@@ -194,15 +219,16 @@ public class WaitUntil<V> extends MultiPhaseAction {
     }
 
     ActionComposer actionComposer = this.getComposer();
-    FluentWait<WebDriver> wait = new FluentWait<>(actionComposer.getWebDriver())
-        .withTimeout(Duration.ofMillis(phaseTimeout))
-        .pollingEvery(Duration.ofMillis(pollInterval))
-        .ignoreAll(ignoreExceptions);
+    FluentWait<WebDriver> wait =
+        new FluentWait<>(actionComposer.getWebDriver())
+            .withTimeout(Duration.ofMillis(phaseTimeout))
+            .pollingEvery(Duration.ofMillis(pollInterval))
+            .ignoreAll(ignoreExceptions);
 
     V result = null;
     try {
-      //firefox doesn't switch focus to top after switch to window, so recovery step is required
-      switchToTopForFirefox(); 
+      // firefox doesn't switch focus to top after switch to window, so recovery step is required
+      switchToTopForFirefox();
       result = wait.until(conditionFunc);
     } catch (TimeoutException e) {
       if (isTimeout()) {
@@ -211,20 +237,24 @@ public class WaitUntil<V> extends MultiPhaseAction {
       }
     }
 
-    //condition not met
+    // condition not met
     if (result == null || (Boolean.class == result.getClass() && Boolean.FALSE.equals(result))) {
       if (isTimeout()) {
         timeoutToDo();
       }
     } else {
-      //condition met => no next phase
+      // condition met => no next phase
       this.noNextPhase();
     }
   }
 
   @Override
   public String toString() {
-    return String.format("%s:%s/%s/%s", WaitUntil.class.getSimpleName(),
-        String.valueOf(totalTimeout), String.valueOf(phaseTimeout), String.valueOf(pollInterval));
+    return String.format(
+        "%s:%s/%s/%s",
+        WaitUntil.class.getSimpleName(),
+        String.valueOf(totalTimeout),
+        String.valueOf(phaseTimeout),
+        String.valueOf(pollInterval));
   }
 }
