@@ -18,7 +18,7 @@ package org.kquiet.browser.action.exception;
 
 /**
  * Thrown to indicate that a browser action encounters a non-recoverable problem during execution.
- * 
+ *
  * @author Kimberly
  */
 public class ActionException extends RuntimeException {
@@ -26,7 +26,7 @@ public class ActionException extends RuntimeException {
 
   /**
    * Constructs an {@link ActionException} with the specified detail message.
-   * 
+   *
    * @param message the detail message
    */
   public ActionException(String message) {
@@ -35,7 +35,7 @@ public class ActionException extends RuntimeException {
 
   /**
    * Constructs an {@link ActionException} with the specified cause.
-   * 
+   *
    * @param cause the cause
    */
   public ActionException(Throwable cause) {
@@ -44,7 +44,7 @@ public class ActionException extends RuntimeException {
 
   /**
    * Constructs a new exception with the specified detail message and cause.
-   * 
+   *
    * @param message the detail message
    * @param cause the cause
    */
