@@ -25,14 +25,12 @@ import java.util.function.Consumer;
 /**
  * {@link PausableThreadPoolExecutor} which execute tasks by their priority. This class is for
  * internal use.
- * 
+ *
  * @author Kimberly
  */
 public class PausablePriorityThreadPoolExecutor extends PausableThreadPoolExecutor {
 
-  /**
-   * Create a {@link PausablePriorityThreadPoolExecutor} with core/maximum pool size set to one.
-   */
+  /** Create a {@link PausablePriorityThreadPoolExecutor} with core/maximum pool size set to one. */
   public PausablePriorityThreadPoolExecutor() {
     this("", 1, 1);
   }
@@ -40,59 +38,73 @@ public class PausablePriorityThreadPoolExecutor extends PausableThreadPoolExecut
   /**
    * Create a {@link PausablePriorityThreadPoolExecutor} with specified parameters and max pool size
    * to five.
-   * 
+   *
    * @param poolPrefix prefix name of thread pool
    * @param corePoolSize core pool size
    * @param maximumPoolSize maximum pool size
    */
-  public PausablePriorityThreadPoolExecutor(String poolPrefix, int corePoolSize,
-      int maximumPoolSize) {
+  public PausablePriorityThreadPoolExecutor(
+      String poolPrefix, int corePoolSize, int maximumPoolSize) {
     this(poolPrefix, corePoolSize, maximumPoolSize, 5);
   }
 
   /**
    * Create a {@link PausablePriorityThreadPoolExecutor} with specified parameters.
-   * 
+   *
    * @param poolPrefix prefix name of thread pool
    * @param corePoolSize core pool size
    * @param maximumPoolSize maximum pool size
    * @param queueSize the maximum queue size used for holding tasks before they are executed
    */
-  public PausablePriorityThreadPoolExecutor(String poolPrefix, int corePoolSize,
-      int maximumPoolSize, int queueSize) {
+  public PausablePriorityThreadPoolExecutor(
+      String poolPrefix, int corePoolSize, int maximumPoolSize, int queueSize) {
     this(poolPrefix, corePoolSize, maximumPoolSize, queueSize, null);
   }
 
   /**
    * Create a {@link PausablePriorityThreadPoolExecutor} with specified parameters.
-   * 
+   *
    * @param poolPrefix prefix name of thread pool
    * @param corePoolSize core pool size
    * @param maximumPoolSize maximum pool size
    * @param afterExecuteFunc the function to execute after any task is executed
    */
-  public PausablePriorityThreadPoolExecutor(String poolPrefix, int corePoolSize,
-      int maximumPoolSize, Consumer<Runnable> afterExecuteFunc) {
+  public PausablePriorityThreadPoolExecutor(
+      String poolPrefix,
+      int corePoolSize,
+      int maximumPoolSize,
+      Consumer<Runnable> afterExecuteFunc) {
     this(poolPrefix, corePoolSize, maximumPoolSize, 5, afterExecuteFunc);
   }
 
   /**
    * Create a {@link PausablePriorityThreadPoolExecutor} with specified parameters.
+   *
    * @param poolPrefix prefix name of thread pool
    * @param corePoolSize core pool size
    * @param maximumPoolSize maximum pool size
    * @param queueSize the maximum queue size used for holding tasks before they are executed
    * @param afterExecuteFunc the function to execute after any task is executed
    */
-  public PausablePriorityThreadPoolExecutor(String poolPrefix, int corePoolSize,
-      int maximumPoolSize, int queueSize, Consumer<Runnable> afterExecuteFunc) {
-    this(poolPrefix, corePoolSize, maximumPoolSize, queueSize, 10L, TimeUnit.MINUTES,
+  public PausablePriorityThreadPoolExecutor(
+      String poolPrefix,
+      int corePoolSize,
+      int maximumPoolSize,
+      int queueSize,
+      Consumer<Runnable> afterExecuteFunc) {
+    this(
+        poolPrefix,
+        corePoolSize,
+        maximumPoolSize,
+        queueSize,
+        10L,
+        TimeUnit.MINUTES,
         afterExecuteFunc);
   }
 
   /**
    * Create a {@link PausablePriorityThreadPoolExecutor} with specified parameters.
-   * 
+   *
    * @param poolPrefix prefix name of thread pool
    * @param corePoolSize core pool size
    * @param maximumPoolSize maximum pool size
@@ -102,12 +114,23 @@ public class PausablePriorityThreadPoolExecutor extends PausableThreadPoolExecut
    * @param unit the time unit for the {@code keepAliveTime} argument
    * @param afterExecuteFunc the function to execute after any task is executed
    */
-  public PausablePriorityThreadPoolExecutor(String poolPrefix, int corePoolSize,
-      int maximumPoolSize, int queueSize, long keepAliveTime, TimeUnit unit,
+  public PausablePriorityThreadPoolExecutor(
+      String poolPrefix,
+      int corePoolSize,
+      int maximumPoolSize,
+      int queueSize,
+      long keepAliveTime,
+      TimeUnit unit,
       Consumer<Runnable> afterExecuteFunc) {
-    super(poolPrefix, corePoolSize, maximumPoolSize, keepAliveTime, unit,
+    super(
+        poolPrefix,
+        corePoolSize,
+        maximumPoolSize,
+        keepAliveTime,
+        unit,
         new PriorityBlockingQueue<Runnable>(queueSize, new PriorityRunnableFutureComparator()),
-        new CommonThreadFactory(poolPrefix), afterExecuteFunc);
+        new CommonThreadFactory(poolPrefix),
+        afterExecuteFunc);
   }
 
   @Override
