@@ -19,14 +19,14 @@ package org.kquiet.concurrent;
 /**
  * {@link Prioritized} is for internal use and should be implemented by the classes whose instances
  * are intended to be executed by {@link PausablePriorityThreadPoolExecutor}.
- * 
+ *
  * @author Kimberly
  */
 public interface Prioritized {
 
   /**
    * Get priority. Smaller value means higher priority.
-   * 
+   *
    * @return priority
    */
   int getPriority();
