@@ -41,30 +41,25 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
-
 import org.kquiet.browser.ActionComposer;
 import org.kquiet.browser.ActionComposerBuilder;
 import org.kquiet.browser.ActionRunner;
 import org.kquiet.browser.action.Custom;
 import org.kquiet.browser.action.ReplyAlert;
-
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import org.takes.Request;
 import org.takes.facets.fork.FkRegex;
 import org.takes.facets.fork.TkFork;
@@ -78,7 +73,7 @@ import org.takes.rs.RsWithType;
 
 /**
  * Test supported actions.
- * 
+ *
  * @author Kimberly
  */
 public class ActionTest {
@@ -93,12 +88,9 @@ public class ActionTest {
   private static URL framesetHtmlFileUrl;
   private static URL innerHtmlFileUrl;
 
-  public ActionTest() {
-  }
+  public ActionTest() {}
 
-  /**
-   * Set up a browser, resources files, and a web server for testing.
-   */
+  /** Set up a browser, resources files, and a web server for testing. */
   @BeforeAll
   public static void setUpClass() {
     browserRunner = TestHelper.createRunner(5);
@@ -108,9 +100,7 @@ public class ActionTest {
     createTempWebServer(TEST_SERVER_EXIT_FLAG);
   }
 
-  /**
-   * Release various resources.
-   */
+  /** Release various resources. */
   @AfterAll
   public static void tearDownClass() {
     browserRunner.close();
@@ -118,12 +108,10 @@ public class ActionTest {
   }
 
   @BeforeEach
-  public void setUp(){
-  }
+  public void setUp() {}
 
   @AfterEach
-  public void tearDown() {
-  }
+  public void tearDown() {}
 
   private static List<By> getFrameSequence(boolean withInFrame) {
     if (withInFrame) {
@@ -133,8 +121,8 @@ public class ActionTest {
     }
   }
 
-  private static ExpectedCondition<?> mergeFrameExpectedCondition(List<By> frameSequence,
-      ExpectedCondition<?> condition) {
+  private static ExpectedCondition<?> mergeFrameExpectedCondition(
+      List<By> frameSequence, ExpectedCondition<?> condition) {
     if (frameSequence.size() == 2) {
       return ExpectedConditions.and(
           ExpectedConditions.frameToBeAvailableAndSwitchToIt(frameSequence.get(0)),
@@ -150,65 +138,72 @@ public class ActionTest {
   }
 
   private static FkRegex getStaticHtml(String pattern, URL fileUrl) {
-    return new FkRegex(pattern, 
+    return new FkRegex(
+        pattern,
         (Request req) -> {
           RqGreedy cachedReq = new RqGreedy(req);
           String method = new RqMethod.Base(cachedReq).method();
           if (RqMethod.GET.equalsIgnoreCase(method)) {
             return new RsWithStatus(
                 new RsWithType(
-                    new RsWithBody(new String(Files.readAllBytes(
-                        Paths.get(new File(fileUrl.getFile()).getAbsolutePath())))), "text/html"),
+                    new RsWithBody(
+                        new String(
+                            Files.readAllBytes(
+                                Paths.get(new File(fileUrl.getFile()).getAbsolutePath())))),
+                    "text/html"),
                 200);
           } else {
             return new RsWithStatus(new RsWithBody("method is not GET:" + method), 200);
           }
-        }
-    );
+        });
   }
 
-  //create an temporary web server for testing
+  // create an temporary web server for testing
   private static void createTempWebServer(AtomicBoolean exitFlag) {
-    new Thread(() -> {
-      try {
-        new FtBasic(
-            new TkFork(
-                new FkRegex("/postform", 
-                    (Request req) -> {
-                      RqGreedy cachedReq = new RqGreedy(req);
-                      String method = new RqMethod.Base(cachedReq).method();
-                      if (RqMethod.POST.equalsIgnoreCase(method)) {
-                        RqFormSmart rfs = new RqFormSmart(cachedReq);
-                        String value1 = rfs.single("key1");
-                        String value2 = rfs.single("key2");
-                        return new RsWithStatus(
-                            new RsWithType(
-                                new RsWithBody("<html>" + value1 + ":" + value2 + "</html>"),
-                                  "text/html"),
-                            200);
-                      } else {
-                        return new RsWithStatus(new RsWithBody("method is not POST:" + method),
-                            200);
-                      }
-                    }
-                ),
-                getStaticHtml("/" + ENTRY_HTML_FILE_NAME, entryHtmlFileUrl),
-                getStaticHtml("/" + FRAMESET_HTML_FILE_NAME, framesetHtmlFileUrl),
-                getStaticHtml("/" + INNER_HTML_FILE_NAME, innerHtmlFileUrl)
-            ), 62226
-        ).start(() -> {
-          return exitFlag.get();
-        });
-      } catch (IOException ex) {
-        throw new RuntimeException(ex);
-      }
-    }).start();
+    new Thread(
+            () -> {
+              try {
+                new FtBasic(
+                        new TkFork(
+                            new FkRegex(
+                                "/postform",
+                                (Request req) -> {
+                                  RqGreedy cachedReq = new RqGreedy(req);
+                                  String method = new RqMethod.Base(cachedReq).method();
+                                  if (RqMethod.POST.equalsIgnoreCase(method)) {
+                                    RqFormSmart rfs = new RqFormSmart(cachedReq);
+                                    String value1 = rfs.single("key1");
+                                    String value2 = rfs.single("key2");
+                                    return new RsWithStatus(
+                                        new RsWithType(
+                                            new RsWithBody(
+                                                "<html>" + value1 + ":" + value2 + "</html>"),
+                                            "text/html"),
+                                        200);
+                                  } else {
+                                    return new RsWithStatus(
+                                        new RsWithBody("method is not POST:" + method), 200);
+                                  }
+                                }),
+                            getStaticHtml("/" + ENTRY_HTML_FILE_NAME, entryHtmlFileUrl),
+                            getStaticHtml("/" + FRAMESET_HTML_FILE_NAME, framesetHtmlFileUrl),
+                            getStaticHtml("/" + INNER_HTML_FILE_NAME, innerHtmlFileUrl)),
+                        62226)
+                    .start(
+                        () -> {
+                          return exitFlag.get();
+                        });
+              } catch (IOException ex) {
+                throw new RuntimeException(ex);
+              }
+            })
+        .start();
   }
 
   private ActionComposerBuilder getDefinedActionComposerBuilder() {
     return new ActionComposerBuilder()
         .prepareActionSequence()
-          .getUrl("http://127.0.0.1:62226/ActionTest.html")
+        .getUrl("http://127.0.0.1:62226/ActionTest.html")
         .returnToComposerBuilder();
   }
 
@@ -217,199 +212,244 @@ public class ActionTest {
   }
 
   @Test
-  public void waitUntilSuccess() {
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(driver -> true, 500)
-        .returnToComposerBuilder()
-        .buildBasic("waitUntilSuccess")
-        .setOpenWindow(false).setCloseWindow(false);
+  void waitUntilSuccess() {
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .waitUntil(driver -> true, 500)
+            .returnToComposerBuilder()
+            .buildBasic("waitUntilSuccess")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(1000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(1000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void waitUntilTimeout() {
+  void waitUntilTimeout() {
     AtomicBoolean result = new AtomicBoolean(false);
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .prepareWaitUntil(driver -> false, 300)
-            .withTimeoutCallback(ac -> {
-              result.set(true);
-            }).done()
-        .returnToComposerBuilder()
-        .buildBasic("waitUntilTimeout")
-        .setOpenWindow(false).setCloseWindow(false).keepFailInfo(false);
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .prepareWaitUntil(driver -> false, 300)
+            .withTimeoutCallback(
+                ac -> {
+                  result.set(true);
+                })
+            .done()
+            .returnToComposerBuilder()
+            .buildBasic("waitUntilTimeout")
+            .setOpenWindow(false)
+            .setCloseWindow(false)
+            .keepFailInfo(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(600, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () -> browserRunner.executeComposer(actionComposer).get(600, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"),
         () -> assertTrue(result.get(), "timeout callback not fired"));
   }
 
   @Test
-  public void justWait() {
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .justWait(300)
-        .returnToComposerBuilder()
-        .buildBasic("justWait")
-        .setOpenWindow(false).setCloseWindow(false);
+  void justWait() {
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .justWait(300)
+            .returnToComposerBuilder()
+            .buildBasic("justWait")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(600, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () -> browserRunner.executeComposer(actionComposer).get(600, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void custom() {
+  void custom() {
     AtomicReference<String> text = new AtomicReference<>("original");
     String customText = "custom";
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .custom(ac -> text.set(customText))
-        .returnToComposerBuilder()
-        .buildBasic("custom")
-        .setOpenWindow(false).setCloseWindow(false);
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .custom(ac -> text.set(customText))
+            .returnToComposerBuilder()
+            .buildBasic("custom")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(1000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(1000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertEquals(customText, text.get()),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void ifThenElse() {
+  void ifThenElse() {
     StringBuilder sb = new StringBuilder();
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .prepareIfThenElse(s -> true)
-          .then()
-            .custom(ac -> {
-              sb.append("1");
-              ac.addToTailByContext(
-                  new Custom((Consumer<ActionComposer>)iac -> sb.append("A"), null));
-              ac.addToTail(
-                  new Custom((Consumer<ActionComposer>)iac -> sb.append("D"), null));
-            })
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .prepareIfThenElse(s -> true)
+            .then()
+            .custom(
+                ac -> {
+                  sb.append("1");
+                  ac.addToTailByContext(
+                      new Custom((Consumer<ActionComposer>) iac -> sb.append("A"), null));
+                  ac.addToTail(new Custom((Consumer<ActionComposer>) iac -> sb.append("D"), null));
+                })
             .prepareIfThenElse(s -> false)
             .then()
-              .custom(ac -> sb.append("2"))
-              .endActionSequence()
+            .custom(ac -> sb.append("2"))
+            .endActionSequence()
             .otherwise()
-              .prepareIfThenElse(s -> true)
-              .then()
-                .custom(ac -> {
+            .prepareIfThenElse(s -> true)
+            .then()
+            .custom(
+                ac -> {
                   sb.append("3");
                   ac.addToPositionByContext(
-                      new Custom((Consumer<ActionComposer>)iac -> sb.append("B"), null), 1);
+                      new Custom((Consumer<ActionComposer>) iac -> sb.append("B"), null), 1);
                   ac.addToPosition(
-                      new Custom((Consumer<ActionComposer>)iac -> sb.append("E"), null), 1);
+                      new Custom((Consumer<ActionComposer>) iac -> sb.append("E"), null), 1);
                 })
-                .custom(ac -> sb.append("C"))
-                .endActionSequence()
-              .endIf()
-              .custom(ac -> sb.append("4"))
-              .endActionSequence()
+            .custom(ac -> sb.append("C"))
+            .endActionSequence()
+            .endIf()
+            .custom(ac -> sb.append("4"))
+            .endActionSequence()
             .endIf()
             .endActionSequence()
-          .otherwise()
+            .otherwise()
             .custom(ac -> sb.append("5"))
             .prepareIfThenElse(s -> true)
             .then()
-              .custom(ac -> sb.append("6"))
-              .endActionSequence()
+            .custom(ac -> sb.append("6"))
+            .endActionSequence()
             .otherwise()
-              .custom(ac -> sb.append("7"))
-              .endActionSequence()
+            .custom(ac -> sb.append("7"))
+            .endActionSequence()
             .endIf()
             .endActionSequence()
-          .endIf()
-        .returnToComposerBuilder()
-        .buildBasic("ifThenElse")
-        .setOpenWindow(false).setCloseWindow(false);
+            .endIf()
+            .returnToComposerBuilder()
+            .buildBasic("ifThenElse")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertEquals("13BC4AED", sb.toString(), "condition sequence not match"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void getUrl() {
-    ActionComposer actionComposer = getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(ExpectedConditions.titleIs("ActionTest"), 3000)
-        .returnToComposerBuilder()  
-        .buildBasic("getUrl")
-        .setOpenWindow(true).setCloseWindow(true);
+  void getUrl() {
+    ActionComposer actionComposer =
+        getDefinedActionComposerBuilder()
+            .prepareActionSequence()
+            .waitUntil(ExpectedConditions.titleIs("ActionTest"), 3000)
+            .returnToComposerBuilder()
+            .buildBasic("getUrl")
+            .setOpenWindow(true)
+            .setCloseWindow(true);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(4000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(4000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void postForm() {
+  void postForm() {
     List<SimpleImmutableEntry<String, String>> formData = new ArrayList<>();
     formData.add(new SimpleImmutableEntry<>("key1", "value1"));
     formData.add(new SimpleImmutableEntry<>("key2", "value2"));
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .postForm("http://127.0.0.1:62226/postform", formData)
-          .waitUntil(ExpectedConditions.and(
-            ExpectedConditions.urlContains("postform"),
-            ExpectedConditions.textToBePresentInElementLocated(By.tagName("html"), "value1:value2")
-            ), 1000)
-        .returnToComposerBuilder()
-        .buildBasic("postForm")
-        .setOpenWindow(true).setCloseWindow(true);
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .postForm("http://127.0.0.1:62226/postform", formData)
+            .waitUntil(
+                ExpectedConditions.and(
+                    ExpectedConditions.urlContains("postform"),
+                    ExpectedConditions.textToBePresentInElementLocated(
+                        By.tagName("html"), "value1:value2")),
+                1000)
+            .returnToComposerBuilder()
+            .buildBasic("postForm")
+            .setOpenWindow(true)
+            .setCloseWindow(true);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void click() {
+  void click() {
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) ->
-        getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-            ExpectedConditions.visibilityOfElementLocated(By.id("btnClick"))), 3000)
-          .prepareClick(By.id("btnClick")).withInFrame(frameSequence.get()).done()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-            ExpectedConditions.visibilityOfElementLocated(By.id("divClickResult"))), 1000)
-        .returnToComposerBuilder()
-        .buildBasic("click" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfElementLocated(By.id("btnClick"))),
+                    3000)
+                .prepareClick(By.id("btnClick"))
+                .withInFrame(frameSequence.get())
+                .done()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfElementLocated(By.id("divClickResult"))),
+                    1000)
+                .returnToComposerBuilder()
+                .buildBasic("click" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time" + frameIdentifier(flag)),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time" + frameIdentifier(flag)),
+                () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       frameSequence.set(getFrameSequence(withFrame));
       assertConsumer.accept(withFrame, composerSupplier.apply(withFrame));
@@ -417,30 +457,47 @@ public class ActionTest {
   }
 
   @Test
-  public void mouseOver() {
+  void mouseOver() {
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) -> getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(), ExpectedConditions.and(
-              ExpectedConditions.visibilityOfElementLocated(By.id("divMouseOver")),
-              ExpectedConditions.invisibilityOfElementLocated(By.id("spMouseOver")))), 3000)
-          .prepareScrollToView(By.id("divMouseOver"), true).withInFrame(frameSequence.get()).done()
-          .prepareMouseOver(By.id("divMouseOver")).withInFrame(frameSequence.get()).done()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.visibilityOfElementLocated(By.id("spMouseOver"))), 1000)
-        .returnToComposerBuilder()
-        .buildBasic("mouseOver" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.and(
+                            ExpectedConditions.visibilityOfElementLocated(By.id("divMouseOver")),
+                            ExpectedConditions.invisibilityOfElementLocated(By.id("spMouseOver")))),
+                    3000)
+                .prepareScrollToView(By.id("divMouseOver"), true)
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareMouseOver(By.id("divMouseOver"))
+                .withInFrame(frameSequence.get())
+                .done()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfElementLocated(By.id("spMouseOver"))),
+                    1000)
+                .returnToComposerBuilder()
+                .buildBasic("mouseOver" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time" + frameIdentifier(flag)),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time" + frameIdentifier(flag)),
+                () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       frameSequence.set(getFrameSequence(withFrame));
       assertConsumer.accept(withFrame, composerSupplier.apply(withFrame));
@@ -448,66 +505,103 @@ public class ActionTest {
   }
 
   @Test
-  public void sendKeys() {
+  void sendKeys() {
     final AtomicReference<String> actualValue = new AtomicReference<>();
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) -> getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.visibilityOfElementLocated(By.id("txtSendKey"))), 3000)
-          .prepareSendKey(By.id("txtSendKey"), "clearbeforesend")
-            .withClearBeforeSend().withInFrame(frameSequence.get()).done()
-          .prepareWaitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.attributeToBe(
-                  By.id("txtSendKey"), "value", "clearbeforesend")), 1000)
-            .withTimeoutCallback(ac -> {
-              try {
-                WebDriver driver = ac.getWebDriver();
-                WebElement inputE = driver.findElements(
-                    By.id("txtSendKey")).stream().findFirst().orElse(null);
-                actualValue.set(Optional.ofNullable(
-                    inputE == null ? null : inputE.getAttribute("value")).orElse(""));
-                ac.skipToFail();
-              } catch (Exception ex) {
-                LOGGER.trace("sendKey error", ex);
-                throw ex;
-              }
-            })
-            .done()
-          .justWait(100)
-          .prepareSendKey(By.id("txtSendKey"), "sendwithoutclear")
-            .withInFrame(frameSequence.get()).done()
-          .prepareWaitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.attributeToBe(
-                  By.id("txtSendKey"), "value", "clearbeforesendsendwithoutclear")), 1000)
-            .withTimeoutCallback(ac -> {
-              try {
-                WebDriver driver = ac.getWebDriver();
-                WebElement inputE = driver.findElements(By.id("txtSendKey"))
-                    .stream().findFirst().orElse(null);
-                actualValue.set(Optional.ofNullable(
-                    inputE == null ? null : inputE.getAttribute("value")).orElse(""));
-                ac.skipToFail();
-              } catch (Exception ex) {
-                LOGGER.trace("sendKey error", ex);
-                throw ex;
-              }
-            })
-            .done()
-        .returnToComposerBuilder()
-        .buildBasic("sendKeys" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfElementLocated(By.id("txtSendKey"))),
+                    3000)
+                .prepareSendKey(By.id("txtSendKey"), "clearbeforesend")
+                .withClearBeforeSend()
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareWaitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.attributeToBe(
+                            By.id("txtSendKey"), "value", "clearbeforesend")),
+                    1000)
+                .withTimeoutCallback(
+                    ac -> {
+                      try {
+                        WebDriver driver = ac.getWebDriver();
+                        WebElement inputE =
+                            driver
+                                .findElements(By.id("txtSendKey"))
+                                .stream()
+                                .findFirst()
+                                .orElse(null);
+                        actualValue.set(
+                            Optional.ofNullable(
+                                    inputE == null ? null : inputE.getAttribute("value"))
+                                .orElse(""));
+                        ac.skipToFail();
+                      } catch (Exception ex) {
+                        LOGGER.trace("sendKey error", ex);
+                        throw ex;
+                      }
+                    })
+                .done()
+                .justWait(100)
+                .prepareSendKey(By.id("txtSendKey"), "sendwithoutclear")
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareWaitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.attributeToBe(
+                            By.id("txtSendKey"), "value", "clearbeforesendsendwithoutclear")),
+                    1000)
+                .withTimeoutCallback(
+                    ac -> {
+                      try {
+                        WebDriver driver = ac.getWebDriver();
+                        WebElement inputE =
+                            driver
+                                .findElements(By.id("txtSendKey"))
+                                .stream()
+                                .findFirst()
+                                .orElse(null);
+                        actualValue.set(
+                            Optional.ofNullable(
+                                    inputE == null ? null : inputE.getAttribute("value"))
+                                .orElse(""));
+                        ac.skipToFail();
+                      } catch (Exception ex) {
+                        LOGGER.trace("sendKey error", ex);
+                        throw ex;
+                      }
+                    })
+                .done()
+                .returnToComposerBuilder()
+                .buildBasic("sendKeys" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time" + frameIdentifier(flag)),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail,"
-            + "actualValue:" + actualValue.get() + frameIdentifier(flag)));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time" + frameIdentifier(flag)),
+                () ->
+                    assertTrue(
+                        ac.isSuccessfulDone(),
+                        "composer fail,"
+                            + "actualValue:"
+                            + actualValue.get()
+                            + frameIdentifier(flag)));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       actualValue.set("original");
       frameSequence.set(getFrameSequence(withFrame));
@@ -516,45 +610,72 @@ public class ActionTest {
   }
 
   @Test
-  public void select() {
+  void select() {
     final AtomicReference<String> actualValue = new AtomicReference<>();
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) -> getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.visibilityOfElementLocated(By.id("slcSelect"))), 3000)
-          .prepareSelect(By.id("slcSelect")).selectByText("option2")
-            .withInFrame(frameSequence.get()).done()
-          .prepareWaitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.attributeToBe(By.id("slcSelect"), "value", "optionValue2")), 1000)
-            .withTimeoutCallback(ac -> {
-              try {
-                WebDriver driver = ac.getWebDriver();
-                WebElement inputE = driver.findElements(By.id("slcSelect"))
-                    .stream().findFirst().orElse(null);
-                actualValue.set(Optional.ofNullable(
-                    inputE == null ? null : inputE.getAttribute("value")).orElse(""));
-                ac.skipToFail();
-              } catch (Exception ex) {
-                LOGGER.trace("select error", ex);
-                throw ex;
-              }
-            })
-            .done()
-        .returnToComposerBuilder()
-        .buildBasic("select" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfElementLocated(By.id("slcSelect"))),
+                    3000)
+                .prepareSelect(By.id("slcSelect"))
+                .selectByText("option2")
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareWaitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.attributeToBe(
+                            By.id("slcSelect"), "value", "optionValue2")),
+                    1000)
+                .withTimeoutCallback(
+                    ac -> {
+                      try {
+                        WebDriver driver = ac.getWebDriver();
+                        WebElement inputE =
+                            driver
+                                .findElements(By.id("slcSelect"))
+                                .stream()
+                                .findFirst()
+                                .orElse(null);
+                        actualValue.set(
+                            Optional.ofNullable(
+                                    inputE == null ? null : inputE.getAttribute("value"))
+                                .orElse(""));
+                        ac.skipToFail();
+                      } catch (Exception ex) {
+                        LOGGER.trace("select error", ex);
+                        throw ex;
+                      }
+                    })
+                .done()
+                .returnToComposerBuilder()
+                .buildBasic("select" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time" + frameIdentifier(flag)),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail,"
-            + "actualValue:" + actualValue.get() + frameIdentifier(flag)));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time" + frameIdentifier(flag)),
+                () ->
+                    assertTrue(
+                        ac.isSuccessfulDone(),
+                        "composer fail,"
+                            + "actualValue:"
+                            + actualValue.get()
+                            + frameIdentifier(flag)));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       actualValue.set("original");
       frameSequence.set(getFrameSequence(withFrame));
@@ -563,73 +684,124 @@ public class ActionTest {
   }
 
   @Test
-  public void openAndCloseWindow() {
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .prepareOpenWindow().registerAs("newwindow").done()
-          .custom(ac -> {
-            if (!ac.switchToWindow(ac.getRegisteredWindow("newwindow"))) {
-              ac.skipToFail();
-            }
-          })
-          .closeWindow(true)
-          .custom(ac -> {
-            if (ac.getWebDriver().getWindowHandles()
-                .contains(ac.getRegisteredWindow("newwindow"))) {
-              ac.skipToFail();
-            }
-          })
-        .returnToComposerBuilder()
-        .buildBasic("openAndCloseWindow")
-        .setOpenWindow(false).setCloseWindow(false);
+  void openAndCloseWindow() {
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .prepareOpenWindow()
+            .registerAs("newwindow")
+            .done()
+            .custom(
+                ac -> {
+                  if (!ac.switchToWindow(ac.getRegisteredWindow("newwindow"))) {
+                    ac.skipToFail();
+                  }
+                })
+            .closeWindow(true)
+            .custom(
+                ac -> {
+                  if (ac.getWebDriver()
+                      .getWindowHandles()
+                      .contains(ac.getRegisteredWindow("newwindow"))) {
+                    ac.skipToFail();
+                  }
+                })
+            .returnToComposerBuilder()
+            .buildBasic("openAndCloseWindow")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void scrollToView() {
+  void scrollToView() {
     final AtomicLong initPosition = new AtomicLong(-1);
     final AtomicLong afterPosition = new AtomicLong(-1);
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) -> getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.visibilityOfAllElementsLocatedBy(By.id("divScroll"))), 1000)
-          .prepareCustom(ac -> {
-            WebElement element = ac.getWebDriver().findElements(By.id("divScroll"))
-                .stream().findFirst().orElse(null);
-            Long top = (Long)((JavascriptExecutor) ac.getWebDriver())
-                .executeScript("return arguments[0].getBoundingClientRect().top;", element);
-            initPosition.set(top);
-          }).withInFrame(frameSequence.get()).done()
-          .prepareScrollToView(By.id("divScroll"), true).withInFrame(frameSequence.get()).done()
-          .prepareCustom(ac -> {
-            WebElement element = ac.getWebDriver().findElements(By.id("divScroll"))
-                .stream().findFirst().orElse(null);
-            Long top = (Long)((JavascriptExecutor) ac.getWebDriver())
-                .executeScript("return arguments[0].getBoundingClientRect().top;", element);
-            afterPosition.set(top);
-          }).withInFrame(frameSequence.get()).done()
-        .returnToComposerBuilder()
-        .buildBasic("scrollToView" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfAllElementsLocatedBy(By.id("divScroll"))),
+                    1000)
+                .prepareCustom(
+                    ac -> {
+                      WebElement element =
+                          ac.getWebDriver()
+                              .findElements(By.id("divScroll"))
+                              .stream()
+                              .findFirst()
+                              .orElse(null);
+                      Object tmp = ((JavascriptExecutor) ac.getWebDriver())
+                          .executeScript(
+                              "return arguments[0].getBoundingClientRect().top;", element);
+                      Long top = -1L;
+                      if (tmp instanceof Double d) {
+                        top = d.longValue();
+                      } else if (tmp instanceof Long l) {
+                        top = l;
+                      }
+                      initPosition.set(top);
+                    })
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareScrollToView(By.id("divScroll"), true)
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareCustom(
+                    ac -> {
+                      WebElement element =
+                          ac.getWebDriver()
+                              .findElements(By.id("divScroll"))
+                              .stream()
+                              .findFirst()
+                              .orElse(null);
+                      Object tmp = ((JavascriptExecutor) ac.getWebDriver())
+                          .executeScript(
+                              "return arguments[0].getBoundingClientRect().top;", element);
+                      Long top = -1L;
+                      if (tmp instanceof Double d) {
+                        top = d.longValue();
+                      } else if (tmp instanceof Long l) {
+                        top = l;
+                      }
+                      afterPosition.set(top);
+                    })
+                .withInFrame(frameSequence.get())
+                .done()
+                .returnToComposerBuilder()
+                .buildBasic("scrollToView" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
-        () -> assertTrue(initPosition.get() > 0 && afterPosition.get() == 0,
-            String.format("scroll error, initial:%s, after:%s",
-                initPosition.get(), afterPosition.get())),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail"));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time"),
+                () ->
+                    assertTrue(
+                        initPosition.get() > 0 && afterPosition.get() == 0,
+                        String.format(
+                            "scroll error, initial:%s, after:%s",
+                            initPosition.get(), afterPosition.get())),
+                () -> assertTrue(ac.isSuccessfulDone(), "composer fail"));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       initPosition.set(-1);
       afterPosition.set(-1);
@@ -639,40 +811,65 @@ public class ActionTest {
   }
 
   @Test
-  public void upload() {
+  void upload() {
     final AtomicReference<String> fileName = new AtomicReference<>();
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) -> getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-              ExpectedConditions.presenceOfElementLocated(By.id("flUpload"))), 1000)
-          .prepareUpload(By.id("flUpload"),
-              new File(entryHtmlFileUrl.getFile()).getAbsolutePath())
-              .withInFrame(frameSequence.get()).done()
-              .waitUntil(mergeFrameExpectedCondition(frameSequence.get(),
-                  ExpectedConditions.visibilityOfElementLocated(By.id("flUpload"))), 1000)
-          .prepareCustom(ac -> {
-            WebElement element = ac.getWebDriver().findElements(
-                By.id("flUpload")).stream().findFirst().orElse(null);
-            String name = (String)((JavascriptExecutor) ac.getWebDriver())
-                .executeScript("return arguments[0].files[0].name;", element);
-            fileName.set(name);
-          }).withInFrame(frameSequence.get()).done()
-        .returnToComposerBuilder()
-        .buildBasic("upload" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.presenceOfElementLocated(By.id("flUpload"))),
+                    1000)
+                .prepareUpload(
+                    By.id("flUpload"), new File(entryHtmlFileUrl.getFile()).getAbsolutePath())
+                .withInFrame(frameSequence.get())
+                .done()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.visibilityOfElementLocated(By.id("flUpload"))),
+                    1000)
+                .prepareCustom(
+                    ac -> {
+                      WebElement element =
+                          ac.getWebDriver()
+                              .findElements(By.id("flUpload"))
+                              .stream()
+                              .findFirst()
+                              .orElse(null);
+                      String name =
+                          (String)
+                              ((JavascriptExecutor) ac.getWebDriver())
+                                  .executeScript("return arguments[0].files[0].name;", element);
+                      fileName.set(name);
+                    })
+                .withInFrame(frameSequence.get())
+                .done()
+                .returnToComposerBuilder()
+                .buildBasic("upload" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time" + frameIdentifier(flag)),
-        () -> assertEquals(ENTRY_HTML_FILE_NAME, fileName.get(),
-            "filename not match" + frameIdentifier(flag)),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time" + frameIdentifier(flag)),
+                () ->
+                    assertEquals(
+                        ENTRY_HTML_FILE_NAME,
+                        fileName.get(),
+                        "filename not match" + frameIdentifier(flag)),
+                () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       fileName.set("");
       frameSequence.set(getFrameSequence(withFrame));
@@ -681,70 +878,109 @@ public class ActionTest {
   }
 
   @Test
-  public void replyAlert() {
-    ActionComposer actionComposer = getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(ExpectedConditions.and(
-              ExpectedConditions.visibilityOfElementLocated(By.id("btnAlert")),
-              ExpectedConditions.visibilityOfElementLocated(By.id("txtAlert"))), 3000)
-          .click(By.id("btnAlert"))
-          .waitUntil(ExpectedConditions.alertIsPresent(), 1000)
-          .prepareReplyAlert(ReplyAlert.Decision.Accept)
-            .withTextAsVariable("AlertMessage").withKeysToSend("AlertInput").done()
-          .waitUntil(ExpectedConditions.and(
-              ExpectedConditions.not(ExpectedConditions.alertIsPresent()),
-              ExpectedConditions.attributeToBe(By.id("txtAlert"), "value", "AlertInput")), 1000)
-        .returnToComposerBuilder()
-        .buildBasic("replyAlert")
-        .setOpenWindow(true).setCloseWindow(true);
+  void replyAlert() {
+    ActionComposer actionComposer =
+        getDefinedActionComposerBuilder()
+            .prepareActionSequence()
+            .waitUntil(
+                ExpectedConditions.and(
+                    ExpectedConditions.visibilityOfElementLocated(By.id("btnAlert")),
+                    ExpectedConditions.visibilityOfElementLocated(By.id("txtAlert"))),
+                3000)
+            .click(By.id("btnAlert"))
+            .waitUntil(ExpectedConditions.alertIsPresent(), 1000)
+            .prepareReplyAlert(ReplyAlert.Decision.Accept)
+            .withTextAsVariable("AlertMessage")
+            .withKeysToSend("AlertInput")
+            .done()
+            .waitUntil(
+                ExpectedConditions.and(
+                    ExpectedConditions.not(ExpectedConditions.alertIsPresent()),
+                    ExpectedConditions.attributeToBe(By.id("txtAlert"), "value", "AlertInput")),
+                1000)
+            .returnToComposerBuilder()
+            .buildBasic("replyAlert")
+            .setOpenWindow(true)
+            .setCloseWindow(true);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
-        () -> assertEquals("PromptMessage", actionComposer.getVariable("AlertMessage"),
-            "alert message not saved as variable or not equal"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(5000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
+        () ->
+            assertEquals(
+                "PromptMessage",
+                actionComposer.getVariable("AlertMessage"),
+                "alert message not saved as variable or not equal"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void extract() {
+  void extract() {
     final AtomicReference<List<By>> frameSequence = new AtomicReference<>();
 
-    Function<Boolean, ActionComposer> composerSupplier = (flag) -> getDefinedActionComposerBuilder()
-        .prepareActionSequence()
-          .waitUntil(mergeFrameExpectedCondition(frameSequence.get(), ExpectedConditions.and(
-            ExpectedConditions.visibilityOfElementLocated(By.id("divMouseOver")),
-            ExpectedConditions.invisibilityOfElementLocated(By.id("spMouseOver")),
-            ExpectedConditions.visibilityOfElementLocated(By.id("txtSendKey")))), 3000)
-          .prepareExtract(By.id("divMouseOver"))
-            .withTextAsVariable("varMouseOver1" + frameIdentifier(flag))
-            .withInFrame(frameSequence.get()).done()
-          .prepareExtract(By.id("txtSendKey"))
-            .withAttributeAsVariable(
-                Stream.of(new String[][]{
-                  {"id", "varSendKey1" + frameIdentifier(flag)},
-                  {"value", "varSendKey2" + frameIdentifier(flag)}})
-                .collect(Collectors.toMap(s -> s[0], s -> s[1])))
-            .withInFrame(frameSequence.get()).done()
-        .returnToComposerBuilder()
-        .buildBasic("extract" + frameIdentifier(flag))
-        .setOpenWindow(true).setCloseWindow(true);
+    Function<Boolean, ActionComposer> composerSupplier =
+        (flag) ->
+            getDefinedActionComposerBuilder()
+                .prepareActionSequence()
+                .waitUntil(
+                    mergeFrameExpectedCondition(
+                        frameSequence.get(),
+                        ExpectedConditions.and(
+                            ExpectedConditions.visibilityOfElementLocated(By.id("divMouseOver")),
+                            ExpectedConditions.invisibilityOfElementLocated(By.id("spMouseOver")),
+                            ExpectedConditions.visibilityOfElementLocated(By.id("txtSendKey")))),
+                    3000)
+                .prepareExtract(By.id("divMouseOver"))
+                .withTextAsVariable("varMouseOver1" + frameIdentifier(flag))
+                .withInFrame(frameSequence.get())
+                .done()
+                .prepareExtract(By.id("txtSendKey"))
+                .withAttributeAsVariable(
+                    Stream.of(
+                            new String[][] {
+                              {"id", "varSendKey1" + frameIdentifier(flag)},
+                              {"value", "varSendKey2" + frameIdentifier(flag)}
+                            })
+                        .collect(Collectors.toMap(s -> s[0], s -> s[1])))
+                .withInFrame(frameSequence.get())
+                .done()
+                .returnToComposerBuilder()
+                .buildBasic("extract" + frameIdentifier(flag))
+                .setOpenWindow(true)
+                .setCloseWindow(true);
 
-    BiConsumer<Boolean, ActionComposer> assertConsumer = (flag, ac) -> assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
-            "not complete in time" + frameIdentifier(flag)),
-        () -> assertEquals("6. mouseover", ac.getVariable("varMouseOver1" + frameIdentifier(flag)),
-            "text of #divMouseOver not saved as variable or not equal" + frameIdentifier(flag)),
-        () -> assertEquals("txtSendKey", ac.getVariable("varSendKey1" + frameIdentifier(flag)),
-            "id of #txtSendKey not saved as variable or not equal" + frameIdentifier(flag)),
-        () -> assertEquals("pre-populated", ac.getVariable("varSendKey2" + frameIdentifier(flag)),
-            "value of #txtSendKey not saved as variable or not equal" + frameIdentifier(flag)),
-        () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
+    BiConsumer<Boolean, ActionComposer> assertConsumer =
+        (flag, ac) ->
+            assertAll(
+                () ->
+                    assertDoesNotThrow(
+                        () -> browserRunner.executeComposer(ac).get(5000, TimeUnit.MILLISECONDS),
+                        "not complete in time" + frameIdentifier(flag)),
+                () ->
+                    assertEquals(
+                        "6. mouseover",
+                        ac.getVariable("varMouseOver1" + frameIdentifier(flag)),
+                        "text of #divMouseOver not saved as variable or not equal"
+                            + frameIdentifier(flag)),
+                () ->
+                    assertEquals(
+                        "txtSendKey",
+                        ac.getVariable("varSendKey1" + frameIdentifier(flag)),
+                        "id of #txtSendKey not saved as variable or not equal"
+                            + frameIdentifier(flag)),
+                () ->
+                    assertEquals(
+                        "pre-populated",
+                        ac.getVariable("varSendKey2" + frameIdentifier(flag)),
+                        "value of #txtSendKey not saved as variable or not equal"
+                            + frameIdentifier(flag)),
+                () -> assertTrue(ac.isSuccessfulDone(), "composer fail" + frameIdentifier(flag)));
 
-    //1:test without frame; 2: test within frame
-    for (int i = 1;i <= 2;i++) {
+    // 1:test without frame; 2: test within frame
+    for (int i = 1; i <= 2; i++) {
       boolean withFrame = (i == 2);
       frameSequence.set(getFrameSequence(withFrame));
       assertConsumer.accept(withFrame, composerSupplier.apply(withFrame));
@@ -752,99 +988,126 @@ public class ActionTest {
   }
 
   @Test
-  public void onFail() {
+  void onFail() {
     StringBuilder sb = new StringBuilder();
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .custom(ac -> {
-            sb.append("custom");
-            throw new RuntimeException("some error");
-          })
-        .returnToComposerBuilder()
-        .onFail(ac -> sb.append("fail"))
-        .onSuccess(ac -> sb.append("success"))
-        .buildBasic("onFail")
-        .setOpenWindow(false).setCloseWindow(false).keepFailInfo(false);
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .custom(
+                ac -> {
+                  sb.append("custom");
+                  throw new RuntimeException("some error");
+                })
+            .returnToComposerBuilder()
+            .onFail(ac -> sb.append("fail"))
+            .onSuccess(ac -> sb.append("success"))
+            .buildBasic("onFail")
+            .setOpenWindow(false)
+            .setCloseWindow(false)
+            .keepFailInfo(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertEquals("customfail", sb.toString(), "on fail not triggered"),
         () -> assertTrue(actionComposer.isFail(), "composer not fail"));
   }
 
   @Test
-  public void onSuccess() {
+  void onSuccess() {
     StringBuilder sb = new StringBuilder();
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .custom(ac -> {
-            sb.append("custom");
-          })
-        .returnToComposerBuilder()
-        .onFail(ac -> sb.append("fail"))
-        .onSuccess(ac -> sb.append("success"))
-        .buildBasic("onSuccess")
-        .setOpenWindow(false).setCloseWindow(false);
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .custom(
+                ac -> {
+                  sb.append("custom");
+                })
+            .returnToComposerBuilder()
+            .onFail(ac -> sb.append("fail"))
+            .onSuccess(ac -> sb.append("success"))
+            .buildBasic("onSuccess")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertEquals("customsuccess", sb.toString(), "on success not triggered"),
         () -> assertTrue(actionComposer.isSuccessfulDone(), "composer fail"));
   }
 
   @Test
-  public void onDone() {
+  void onDone() {
     StringBuilder sb = new StringBuilder();
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .custom(ac -> sb.append("custom"))
-        .returnToComposerBuilder()
-        .onDone(ac -> sb.append("done"))
-        .buildBasic("onDone")
-        .setOpenWindow(false).setCloseWindow(false);
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .custom(ac -> sb.append("custom"))
+            .returnToComposerBuilder()
+            .onDone(ac -> sb.append("done"))
+            .buildBasic("onDone")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
         () -> assertEquals("customdone", sb.toString(), "on done not triggered"),
         () -> assertTrue(actionComposer.isDone(), "composer not done"));
   }
-  
+
   @Test
   @Tag("new")
-  public void actionPerform() {
+  void actionPerform() {
     StringBuilder sb = new StringBuilder();
     AtomicInteger ai = new AtomicInteger(0);
     AtomicInteger ai2 = new AtomicInteger(10);
-    ActionComposer actionComposer = getEmptyActionComposerBuilder()
-        .prepareActionSequence()
-          .justWait(300)
-          .prepareIfThenElse(ac -> true)
-          .then()
-            .custom(ac -> {
-              sb.append("custom");
-            })
+    ActionComposer actionComposer =
+        getEmptyActionComposerBuilder()
+            .prepareActionSequence()
+            .justWait(300)
+            .prepareIfThenElse(ac -> true)
+            .then()
+            .custom(
+                ac -> {
+                  sb.append("custom");
+                })
             .endActionSequence()
-        .returnToComposerBuilder()
-        .actionPerforming(ac -> at -> {
-          sb.append(ai.addAndGet(1));
-        })
-        .actionPerformed(ac -> at -> {
-          sb.append(ai2.addAndGet(-1));
-        })
-        .buildBasic("actionPerforming")
-        .setOpenWindow(false).setCloseWindow(false);
+            .returnToComposerBuilder()
+            .actionPerforming(
+                ac ->
+                    at -> {
+                      sb.append(ai.addAndGet(1));
+                    })
+            .actionPerformed(
+                ac ->
+                    at -> {
+                      sb.append(ai2.addAndGet(-1));
+                    })
+            .buildBasic("actionPerforming")
+            .setOpenWindow(false)
+            .setCloseWindow(false);
 
     assertAll(
-        () -> assertDoesNotThrow(
-            () -> browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
-            "not complete in time"),
-        () -> assertEquals("192384custom76", sb.toString(),
-            "action perform function not triggered corrrectly"));
+        () ->
+            assertDoesNotThrow(
+                () ->
+                    browserRunner.executeComposer(actionComposer).get(3000, TimeUnit.MILLISECONDS),
+                "not complete in time"),
+        () ->
+            assertEquals(
+                "192384custom76",
+                sb.toString(),
+                "action perform function not triggered corrrectly"));
   }
 }
